# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cold Chain Integrated Prototype API",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/control/{device_id}/pending": {
      "get": {
        "tags": [
          "Project compatibility"
        ],
        "summary": "Pending",
        "operationId": "pending_api_v1_control__device_id__pending_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "x-device-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Device-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/control/{command_id}/acknowledge": {
      "post": {
        "tags": [
          "Project compatibility"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_api_v1_control__command_id__acknowledge_post",
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Command Id"
            }
          },
          {
            "name": "x-device-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Device-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandAcknowledgement"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/control/{device_id}/history": {
      "get": {
        "tags": [
          "Project compatibility"
        ],
        "summary": "History",
        "operationId": "history_api_v1_control__device_id__history_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommandAcknowledgement": {
        "properties": {
          "device_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,48}$",
            "title": "Device Id"
          },
          "boot_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Boot Id"
          },
          "sequence": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sequence"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "APPLIED",
              "REJECTED"
            ],
            "title": "Outcome"
          },
          "primary_cooling": {
            "type": "boolean",
            "title": "Primary Cooling"
          },
          "backup_cooling": {
            "type": "boolean",
            "title": "Backup Cooling"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "boot_id",
          "sequence",
          "outcome",
          "primary_cooling",
          "backup_cooling",
          "reason"
        ],
        "title": "CommandAcknowledgement"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}